Split actionable-proposals.derived.ts
